+++ src/client/src/tscSQLParser.c
@@ -1410,7 +1410,7 @@ static int32_t handleArithmeticExpr(SSqlCmd* pCmd, int32_t clauseIndex, int32_t 
       // TODO: other error handling
     } END_TRY
 
-    pInfo->pArithExprInfo->base.arg[0].argBytes = tbufTell(&bw);
+    pInfo->pArithExprInfo->base.arg[0].argBytes = (int16_t) tbufTell(&bw);
     pInfo->pArithExprInfo->base.arg[0].argValue.pz = tbufGetData(&bw, true);
     pInfo->pArithExprInfo->base.arg[0].argType = TSDB_DATA_TYPE_BINARY;
     tbufCloseWriter(&bw);
